Product Get Repository Implements
