+++ src/main/java/com/alpcashierin/apiCalls/pojosModel/productResponse/getProduct/GetProductResp.java
@@ -1,0 +1,20 @@
+package com.alpcashierin.apiCalls.pojosModel.productResponse.getProduct;
+import com.google.gson.annotations.SerializedName;
+public class GetProductResp {
+    @SerializedName("status")
+    String status;
+    @SerializedName("data")
+    ProductResp product;
+
+    public String getStatus() {
+        return status;
+    }
+
+    public ProductResp getProduct() {
+        return product;
+    }
+
+
+}
+
+
